+++ example/lib/avatar_example.dart
@@ -70,55 +70,47 @@ class _AvatarExampleState extends State<AvatarExample> {
 
   List<String> avatar =
       List.generate(52, (index) => "${alphabet[index]} avatar ${index + 1} ");
+
   @override
   Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
     return Scaffold(
       appBar: AppBar(),
-      body: SingleChildScrollView(
-        child: SizedBox(
-          width: size.width,
-          height: size.height,
-          child: Column(
-            children: [
-              /*  ListView.builder(
-                physics: const BouncingScrollPhysics(),
-                itemCount: avatar.length,
-                itemBuilder: (context, index) {
-                  // Check if the index is within the bounds of the list
-                  if (index < avatar.length) {
-                    return Container(
-                      margin: const EdgeInsets.symmetric(vertical: 10),
-                      padding: const EdgeInsets.symmetric(horizontal: 10),
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          Avatar(
-                            onTapAvatar: () {},
-                            radius: 35,
-                            text: avatar[index],
-                            randomGradient: true,
-                            randomColor: false,
-                          ),
-                          Text(avatar[index]),
-                        ],
-                      ),
-                    );
-                  } else {
-                    // Return a placeholder widget if the index is out of bounds
-                    return Container();
-                  }
-                },
-              ),*/
-              const SizedBox(height: 25),
-              Avatar(
-                showPageViewOnTap: true,
-                text: "S",
-                imageNetwork: "https://i.pravatar.cc/300",
-              ),
-            ],
+      body: CustomScrollView(
+        physics: const BouncingScrollPhysics(),
+        slivers: [
+          SliverList(
+            delegate: SliverChildBuilderDelegate(
+              (context, index) {
+                if (index < avatar.length) {
+                  return Container(
+                    margin: const EdgeInsets.symmetric(vertical: 10),
+                    padding: const EdgeInsets.symmetric(horizontal: 10),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        Avatar.profile(
+                          imageNetwork: "https://i.pravatar.cc/300",
+                          // onTapAvatar: () {},
+                          radius: 35,
+                          // showPageViewOnTap: true,
+                          text: avatar[index],
+                          randomGradient: true,
+                          randomColor: false,
+                        ),
+                        Text(avatar[index]),
+                      ],
+                    ),
+                  );
+                } else {
+                  // Return the additional Avatar widget
+                  return const SizedBox(height: 25);
+                }
+              },
+              childCount:
+                  avatar.length + 1, // +1 for the additional Avatar widget
+            ),
           ),
-        ),
+        ],
       ),
     );
   }