+++ CHANGELOG.md
@@ -1,3 +1,45 @@
-## 0.0.1
+## 0.0.5
 
-* TODO: Describe initial release.
+* Fix support web.
+
+## 0.0.6
+
+* Add page view feature.
+
+## 0.0.7
+
+* Update Pack.
+
+## 0.0.8
+
+* Update Pack.
+* Support platform windows and mac.
+* Add new feature.
+
+
+## 0.0.9
+
+* Support all platform cropper image.
+
+## 0.0.9+1
+
+* Adit README.md.
+
+## 0.0.9+2
+
+* Add new feature.
+
+## 0.0.9+3
+
+* Fix color page view background.
+
+## 0.0.9+4
+
+* Fix pub point.
+* Update package
+* Edit README
+* Fix permisstions android
+
+
+
+